Move the font selectors into whitelabel
getFont and getFontFiles derive from the whitelabel font settings, so
they live with the whitelabel module now. styled-components/selectors.ts
is deleted; its importers point at metabase/whitelabel.
