+++ frontend/src/embedding-sdk-bundle/components/private/SdkGlobalFontsStyles.tsx
@@ -4,7 +4,7 @@ import { useMemo } from "react";
 
 import { defaultFontFiles } from "metabase/css/core/fonts.styled";
 import { useSelector } from "metabase/redux";
-import { getFontFiles } from "metabase/styled-components/selectors";
+import { getFontFiles } from "metabase/whitelabel";
 
 /**
  * css style to define the font files for the SDK
