--- frontend/src/metabase/styled-components/selectors.ts
@@ -1,22 +1,0 @@
-import { createSelector } from "@reduxjs/toolkit";
-import _ from "underscore";
-
-import { getEmbedOptions } from "metabase/embedding/interactive-embedding";
-import { getSettings } from "metabase/settings";
-
-export const getFont = createSelector(
-  [getSettings, getEmbedOptions],
-  (settings, embedOptions) => {
-    if (embedOptions.font) {
-      return embedOptions.font;
-    } else if (!_.isEmpty(settings["application-font-files"])) {
-      return "Custom";
-    } else {
-      return settings["application-font"];
-    }
-  },
-);
-
-export const getFontFiles = createSelector([getSettings], (settings) => {
-  return settings["application-font-files"];
-});
